+++ tests/clients/test_ngsi_v2_cb.py
@@ -375,6 +375,7 @@ class TestContextBroker(unittest.TestCase):
             id3 = client.post_subscription(sub2)
             self.assertNotEqual(id1, id3)
 
+
     @clean_test(fiware_service=settings.FIWARE_SERVICE,
                 fiware_servicepath=settings.FIWARE_SERVICEPATH,
                 cb_url=settings.CB_URL,
@@ -805,7 +806,10 @@ class TestContextBroker(unittest.TestCase):
         self.iotac.post_devices(devices=devices)
         while devices:
             device = devices.pop()
-            self.client.delete_entity(entity_id=device.entity_name, entity_type=device.entity_type, delete_devices=True)
+            self.client.delete_entity(entity_id=device.entity_name,
+                                      entity_type=device.entity_type,
+                                      delete_devices=True,
+                                      iota_url=settings.IOTA_URL)
             self.assertEqual(len(self.iotac.get_device_list()), len(devices))
 
     def tearDown(self) -> None: