+++ tests/clients/test_ngsi_v2_iota.py
@@ -21,7 +21,6 @@ from filip.models.ngsi_v2.iot import \
     StaticDeviceAttribute
 from filip.utils.cleanup import clear_all, clean_test
 from tests.config import settings
-from filip.utils.iot import filter_device_list
 
 
 logger = logging.getLogger(__name__)
@@ -165,12 +164,15 @@ class TestAgent(unittest.TestCase):
             logger.info(client.get_entity(entity_id=device.entity_name).json(
                 indent=2))
 
+    @clean_test(fiware_service=settings.FIWARE_SERVICE,
+                fiware_servicepath=settings.FIWARE_SERVICEPATH,
+                cb_url=settings.CB_URL,
+                iota_url=settings.IOTA_JSON_URL)
     def test_deletions(self):
         """
         Test the deletion of a context entity/device if the state is always
         correctly cleared
         """
-        self.tearDown()
 
         device_id = 'device_id'
         entity_id = 'entity_id'
@@ -206,11 +208,14 @@ class TestAgent(unittest.TestCase):
         #        that is linked to multiple devices
         # delete with optional parameter -> entity needs to be deleted
         self.client.post_device(device=device)
+
         device2 = copy.deepcopy(device)
         device2.device_id = "device_id2"
         self.client.post_device(device=device2)
-        self.assertRaises(Exception, self.client.delete_device,
-                          device_id=device_id, delete_entity=True)
+        with self.assertRaises(Exception):
+            self.client.delete_device(device_id=device_id,
+                                      delete_entity=True,
+                                      cb_url=settings.CB_URL)
         self.assertTrue(
             len(cb_client.get_entity_list(entity_ids=[entity_id])) == 1)
         self.client.delete_device(device_id=device2.device_id)
@@ -370,83 +375,6 @@ class TestAgent(unittest.TestCase):
             live_device = self.client.get_device(device_id=device.device_id)
             self.assertEqual(live_device.__getattribute__(key), value)
             cb_client.close()
-
-    def test_filter_device_list(self) -> None:
-        """
-        Test the function filter_device_list
-        """
-        # create devices
-        base_device_id = "device:"
-
-        base_entity_id_1 = "urn:ngsi-ld:TemperatureSensor:"
-        entity_type_1 = "TemperatureSensor"
-
-        base_entity_id_2 = "urn:ngsi-ld:HumiditySensor:"
-        entity_type_2 = "HumiditySensor"
-
-        devices = []
-
-        for i in range(20):
-            base_entity_id = base_entity_id_1 if i < 10 else base_entity_id_2
-            entity_type = entity_type_1 if i < 10 else entity_type_2
-
-            device_id = base_device_id + str(i)
-            entity_id = base_entity_id + str(i)
-            device = Device(
-                service=settings.FIWARE_SERVICE,
-                service_path=settings.FIWARE_SERVICEPATH,
-                device_id=device_id,
-                entity_type=entity_type,
-                entity_name=entity_id
-            )
-            devices.append(device)
-
-        entity_id_list = [device.entity_name for device in devices]
-        device_id_list = [device.device_id for device in devices]
-
-        # test with no inputs
-        self.assertEqual(len(filter_device_list(devices)), len(devices))
-
-        # test with entity type
-        self.assertEqual(len(filter_device_list(devices, entity_types=[entity_type_1])),
-                         10)
-        self.assertEqual(len(filter_device_list(devices, entity_types=[entity_type_1, entity_type_2])),
-                         20)
-
-        # test with entity id
-        self.assertEqual(len(filter_device_list(devices, entity_names=entity_id_list[5:])),
-                         len(entity_id_list[5:]))
-
-        # test with entity type and entity id
-        self.assertEqual(len(filter_device_list(devices, entity_names=entity_id_list, entity_types=[entity_type_1])),
-                         10)
-
-        # test with device id
-        self.assertEqual(len(filter_device_list(devices, device_ids=device_id_list[5:])),
-                         len(device_id_list[5:]))
-
-        # test with single args
-        self.assertEqual(len(filter_device_list(devices,
-                                                device_ids=devices[0].device_id)), 1)
-
-        self.assertEqual(len(filter_device_list(devices,
-                                                entity_names=devices[0].entity_name)), 1)
-
-        self.assertNotEqual(len(filter_device_list(devices,
-                                                   entity_types=devices[0].entity_type)), 1)
-
-        self.assertEqual(len(filter_device_list(devices,
-                                                device_ids=devices[0].device_id,
-                                                entity_names=devices[0].entity_name,
-                                                entity_types=devices[0].entity_type)), 1)
-
-        # test for errors
-        with self.assertRaises(TypeError):
-            filter_device_list(devices, device_ids={'1234'})
-        with self.assertRaises(TypeError):
-            filter_device_list(devices, entity_names={'1234'})
-        with self.assertRaises(TypeError):
-            filter_device_list(devices, entity_types={'1234'})
 
     def tearDown(self) -> None:
         """