+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-#### v0.2.4
-- fixed coupled delete of devices and entities ([#143](https://github.com/RWTH-EBC/FiLiP/issues/143))
-
 #### v0.2.3
+- fixed issue of client generation when multiple clients are required ([#151](https://github.com/RWTH-EBC/FiLiP/issues/151))
+- added filter functions for better subscription handling ([#148](https://github.com/RWTH-EBC/FiLiP/issues/148))
+- added filter functions for better device handling ([#143](https://github.com/RWTH-EBC/FiLiP/issues/143))
+- fixed coupled delete of devices and entities ([#143](https://github.com/RWTH-EBC/FiLiP/issues/143))
 - Add images to tutorials ([#139](https://github.com/RWTH-EBC/FiLiP/issues/139))
 
 #### v0.2.2